requirements.txt
google-adk==1.19.0
google-adk[eval]==1.19.0
pytest==9.0.1
pytest-cov==7.0.0
pytest-asyncio==1.3.0
python-dotenv

